fix: validate custom role permissions before rename
