+++ apps/dokploy/server/api/routers/proprietary/custom-role.ts
@@ -10,6 +10,7 @@ import {
 	protectedProcedure,
 } from "../../trpc";
 import { audit } from "../../utils/audit";
+import { validatePermissions } from "./custom-role-utils";
 
 const permissionsSchema = z.record(z.string(), z.array(z.string()));
 
@@ -150,6 +151,8 @@ export const customRoleRouter = createTRPCRouter({
 				});
 			}
 
+			validatePermissions(input.permissions);
+
 			const effectiveRoleName = input.newRoleName ?? input.roleName;
 
 			if (input.newRoleName && input.newRoleName !== input.roleName) {
@@ -179,8 +182,6 @@ export const customRoleRouter = createTRPCRouter({
 						),
 					);
 			}
-
-			validatePermissions(input.permissions);
 
 			const [updated] = await db
 				.update(organizationRole)
@@ -289,33 +290,3 @@ export const customRoleRouter = createTRPCRouter({
 		return statements;
 	}),
 });
-
-const INTERNAL_RESOURCES = ["organization", "invitation", "team", "ac"];
-
-function validatePermissions(permissions: Record<string, string[]>) {
-	for (const [resource, actions] of Object.entries(permissions)) {
-		if (INTERNAL_RESOURCES.includes(resource)) {
-			throw new TRPCError({
-				code: "BAD_REQUEST",
-				message: `Resource "${resource}" is managed internally and cannot be assigned to custom roles`,
-			});
-		}
-
-		if (!(resource in statements)) {
-			throw new TRPCError({
-				code: "BAD_REQUEST",
-				message: `Unknown resource: ${resource}`,
-			});
-		}
-
-		const validActions = statements[resource as keyof typeof statements];
-		for (const action of actions) {
-			if (!validActions.includes(action as never)) {
-				throw new TRPCError({
-					code: "BAD_REQUEST",
-					message: `Invalid action "${action}" for resource "${resource}". Valid actions: ${validActions.join(", ")}`,
-				});
-			}
-		}
-	}
-}
